+++ docs/examples/page_url_simple.php
@@ -3,6 +3,7 @@ use QATools\QATools\PageObject\Page;
 
 /**
  * @page-url('http://www.example.com/products/shoes.html?color=red')
+ * @match-url-exact('http://www.example.com/products/shoes.html?color=red')
  */
 class NewProductsPage extends Page
 {