+++ tests/QATools/QATools/Live/PageObject/Pages/WebElementPage.php
@@ -15,6 +15,11 @@ use QATools\QATools\PageObject\Element\WebElement;
 use QATools\QATools\PageObject\Page;
 use tests\QATools\QATools\Live\PageObject\Element\InputContainer;
 
+/**
+ * Simple page containing web elements.
+ *
+ * @match-url-component('anchor' => 'test_anchor')
+ */
 class WebElementPage extends Page
 {
 
@@ -66,5 +71,13 @@ class WebElementPage extends Page
 	 * @var InputContainer
 	 */
 	public $inputContainer;
+
+	/**
+	 * Link for anchor match test.
+	 *
+	 * @var WebElement
+	 * @find-by('id' => 'anchor_tester')
+	 */
+	public $linkAnchorTest;
 
 }